Parameter labelled amplitude in GaussianPSF is really the integrated flux
The easy way to see this is to make a GaussianPSF with different parameters for sigma
(see the tests).
Then evaluate the PSF over a grid and sum it up. The sum does not change, but the value of the
central pixel does. So, clearly, this parameter is the total flux, not the amplitude.

The theoretical way to think about it, is that the value of the function is the integrated
flux over a pixel, not the amplitude over a pixel. That's why what's was called `A` in the
formula before is not the amplitude, but the integrated flux.
(and thus I gave it a new character).