Fix deblending with sources not in the segmentation mask
Fix #616.

To deblend a source a small stamp is created and can contain other
sources, depending on the shape of the source mask. These neighboring
sources are not masked so they are used in the deblending process. Given
that the data array is also used for flux computation before the
watershed, the most abvious solution is to put values outside the source
mask to zero.